requirements.txt
aiogram>=3.0
aiohttp>=3.8
python-dotenv>=0.21
Pillow>=9.0
aiogram-calendar==0.6.0
flask>=3.0
flask-cors>=4.0
python-telegram-bot>=20.0